reconfigured the assistant to make it less chonky
